Neith michel

+++ src/main/java/com/projet_QCM/repository/UserRepository.java
@@ -6,8 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.projet_QCM.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer>{
-	
+public interface UserRepository extends JpaRepository<User, Long>{	
 	User findByLogin(String login);
 
 }
